Move common functions to utils/common.py
