+++ adlib/tests/adversaries/data_modification_test.py
@@ -6,8 +6,8 @@
 from adlib.learners import SimpleLearner
 from adlib.adversaries.datamodification.data_modification import \
     DataModification
-from adlib.tests.adversaries.label_flipping_test import \
-    calculate_correct_percentages
+from adlib.utils.common import calculate_correct_percentages
+from adlib.utils.common import get_spam_features
 from copy import deepcopy
 from data_reader.dataset import EmailDataset
 from data_reader.operations import load_dataset
@@ -131,33 +131,6 @@ def test_data_modification():
     print('\nEND data modification attack.')
     print('###################################################################')
     print()
-
-
-def get_spam_features(instances, p=0.9):
-    """
-    Returns a list of feature indices where the proportion of instances that
-    have them is >= p
-    :param instances: the spam instances - MUST BE SPAM (i.e. have a label of 1)
-    :param p: the proportion of instances that must have this value
-    :return: a tuple comprised of spam and ham features in separate lists
-    """
-
-    if len(instances) == 0:
-        raise ValueError('Must have at least one instance.')
-
-    spam_features = []
-    ham_features = []
-    for i in range(instances[0].get_feature_count()):
-        count = 0
-        for inst in instances:
-            count += 1 if inst.get_feature_vector().get_feature(i) > 0 else 0
-
-        if (count / len(instances)) >= p:
-            spam_features.append(i)
-        else:
-            ham_features.append(i)
-
-    return spam_features, ham_features
 
 
 if __name__ == '__main__':
