+++ adlib/utils/common.py
@@ -26,3 +26,62 @@ def logistic_function(x):
     """
 
     return 1 / (1 + math.exp(-1 * x))
+
+
+def get_spam_features(instances, p=0.9):
+    """
+    Returns a list of feature indices where the proportion of instances that
+    have them is >= p
+    :param instances: the spam instances - MUST BE SPAM (i.e. have a label of 1)
+    :param p: the proportion of instances that must have this value
+    :return: a tuple comprised of spam and ham features in separate lists
+    """
+
+    if len(instances) == 0:
+        raise ValueError('Must have at least one instance.')
+
+    spam_features = []
+    ham_features = []
+    for i in range(instances[0].get_feature_count()):
+        count = 0
+        for inst in instances:
+            count += 1 if inst.get_feature_vector().get_feature(i) > 0 else 0
+
+        if (count / len(instances)) >= p:
+            spam_features.append(i)
+        else:
+            ham_features.append(i)
+
+    return spam_features, ham_features
+
+
+def calculate_correct_percentages(orig_labels, attack_labels, instances):
+    """
+    Calculates the percent of labels that were predicted correctly before and
+    after the attack.
+    :param orig_labels: the labels predicted by the pre-attack learner
+    :param attack_labels: the labels predicted by the post-attack learner
+    :param instances: the list of instances
+    :return: strings of original percent correct, attack percent correct, and
+             the difference (original - attack)
+    """
+
+    orig_count = 0
+    count = 0
+    for i in range(len(instances)):
+        if orig_labels[i] != instances[i].get_label():
+            orig_count += 1
+        elif attack_labels[i] != instances[i].get_label():
+            count += 1
+
+    orig_precent_correct = ((len(instances) - orig_count) * 100
+                            / len(instances))
+    attack_precent_correct = ((len(instances) - count) * 100
+                              / len(instances))
+    difference = orig_precent_correct - attack_precent_correct
+
+    orig_precent_correct = str(round(orig_precent_correct, 4))
+    attack_precent_correct = str(round(attack_precent_correct, 4))
+    difference = str(round(difference, 4))
+
+    return orig_precent_correct, attack_precent_correct, difference
